v1.03 Entities and Videos Relationship update
I spent some time cleaning up code and putting what needs to be put into functions away. Less of a mess.

I decided to get rid of the entity_relationships table as I found that  
1. I needed to hold more info that a post_id and an entity_id  
2. I thought it would seem easier and less of a hassle to make one query to a singular database that would hold the information.

There were some bugs introduced whenever you wanted to add a new video post type.   
Whenever you try to make a new post, the `save_hook` method gets called even before you save the video. I guess its a preparation to save.
 Because of this, the `save_post` hook was throwing errors as it couldn't grab information to send to the custom tables I made, because none yet existed. I did find a fix by using one line if statements (e.x. `$thumbnail_id = (isset($thumbnail_id))? $thumbnail_id : null;`), which to my knowledge, have been resolved.

There are new steps in what should be done next. I found that when you create a new video post, but do not save it yet, `save_post` automatically saves it into `cpvm_entities` table as an 'Auto Draft` under the name column. 

Once you change the name and save it, it creates that new entity.  

I need to implement a way that syncs the way posts and entities will interact. When a post is deleted, that entity entry has to be deleted to. If an auto draft 
entry get revised and fixed, it needs to update the entities table. Those 2 tables need to be in constant sync with each other before I can implement the correct API structure.